chore: seperate auth and user routes and add time stamps on user model
